requirements.txt
attrs==21.4.0
autopep8==1.6.0
coverage==6.4.1
flake8==4.0.1
iniconfig==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.29
packaging==21.3
pluggy==1.0.0
protobuf==4.21.1
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
toml==0.10.2
tomli==2.0.1
